Extract vm_lifecylce tests from subresource_api file
