+++ tests/compute/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "console.go",
         "credentials.go",
         "framework.go",
+        "vm_lifecycle.go",
         "vmidefaults.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/compute",
@@ -26,6 +27,7 @@ go_library(
         "//tests/libkubevirt:go_default_library",
         "//tests/libkubevirt/config:go_default_library",
         "//tests/libnet/cloudinit:go_default_library",
+        "//tests/libpod:go_default_library",
         "//tests/libsecret:go_default_library",
         "//tests/libvmifact:go_default_library",
         "//tests/libvmops:go_default_library",
@@ -35,5 +37,6 @@ go_library(
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
     ],
 )
