+++ tests/compute/subresources/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "pausing.go",
         "rbac.go",
         "softreboot.go",
-        "subresource_api.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/compute/subresources",
     visibility = ["//visibility:public"],
@@ -23,11 +22,9 @@ go_library(
         "//staging/src/kubevirt.io/client-go/kubevirt/typed/core/v1:go_default_library",
         "//tests/compute:go_default_library",
         "//tests/console:go_default_library",
-        "//tests/decorators:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libinstancetype/builder:go_default_library",
-        "//tests/libpod:go_default_library",
         "//tests/libvmifact:go_default_library",
         "//tests/libvmops:go_default_library",
         "//tests/libwait:go_default_library",
@@ -40,7 +37,6 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/validation/field:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
     ],
 )
